agents/chat/pyproject.toml
[project]
name = "glideator-chat"
version = "0.1.0"
description = ""
authors = [
    {name = "Jan Helcl"}
]
package-mode = false
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "jupyterlab (>=4.4.6,<5.0.0)",
    "autogen-ext[mcp] (>=0.7.4,<0.8.0)",
    "autogen-agentchat (>=0.7.4,<0.8.0)",
    "autogen-ext[openai] (>=0.7.4,<0.8.0)",
    "gradio (>=5.44.1,<6.0.0)",
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"